+++ Thunderstore/CHANGELOG.md
@@ -1,13 +1,17 @@
 ## 0.1.0
+
 - Publish
 
 ## 0.1.1
+
 - Fix unnecessary dependency error
 
 ## 0.2.0
+
 - Added Convert To configurable property
 - Added Conversion Selection Type configurable property
 - Added Minimum Time configurable property
+- Added Primary Replacement configurable property
 - Migrated Benthic-like transformation to its own properties
 - Adjust default weights for a larger difference and a higher focus on void items
 - Nerf default bomb generation frequency stacking rate (1 -> 0.5)
@@ -15,3 +19,4 @@
 - Nerf default stacking bomb damage (10% -> 5%)
 - Buff default bomb range (15 -> 17.5)
 - Nerf default stacking bomb range (1 -> 0.5)
+- Default multiplier per stack (0.9 -> 1)